chore: update times from i64 to u64

+++ fuse/src/reply.rs
@@ -101,12 +101,12 @@ fn fuse_attr_from_attr(attr: &FileAttr) -> fuse_attr {
         ino: attr.ino,
         size: attr.size,
         blocks: attr.blocks,
-        atime: attr.atime.sec,
-        mtime: attr.mtime.sec,
-        ctime: attr.ctime.sec,
-        atimensec: attr.atime.nsec,
-        mtimensec: attr.mtime.nsec,
-        ctimensec: attr.ctime.nsec,
+        atime: attr.atime.sec as u64,
+        mtime: attr.mtime.sec as u64,
+        ctime: attr.ctime.sec as u64,
+        atimensec: attr.atime.nsec as u32,
+        mtimensec: attr.mtime.nsec as u32,
+        ctimensec: attr.ctime.nsec as u32,
         mode: mode_from_kind_and_perm(attr.kind, attr.perm),
         nlink: attr.nlink,
         uid: attr.uid,
@@ -249,10 +249,10 @@ impl ReplyEntry {
         self.reply.ok(&fuse_entry_out {
             nodeid: attr.ino,
             generation: generation,
-            entry_valid: ttl.sec,
-            attr_valid: ttl.sec,
-            entry_valid_nsec: ttl.nsec,
-            attr_valid_nsec: ttl.nsec,
+            entry_valid: ttl.sec as u64,
+            attr_valid: ttl.sec as u64,
+            entry_valid_nsec: ttl.nsec as u32,
+            attr_valid_nsec: ttl.nsec as u32,
             attr: fuse_attr_from_attr(attr),
         });
     }
@@ -281,8 +281,8 @@ impl ReplyAttr {
     /// Reply to a request with the given attribute
     pub fn attr(self, ttl: &Timespec, attr: &FileAttr) {
         self.reply.ok(&fuse_attr_out {
-            attr_valid: ttl.sec,
-            attr_valid_nsec: ttl.nsec,
+            attr_valid: ttl.sec as u64,
+            attr_valid_nsec: ttl.nsec as u32,
             dummy: 0,
             attr: fuse_attr_from_attr(attr),
         });
@@ -446,10 +446,10 @@ impl ReplyCreate {
         self.reply.ok(&(fuse_entry_out {
             nodeid: attr.ino,
             generation: generation,
-            entry_valid: ttl.sec,
-            attr_valid: ttl.sec,
-            entry_valid_nsec: ttl.nsec,
-            attr_valid_nsec: ttl.nsec,
+            entry_valid: ttl.sec as u64,
+            attr_valid: ttl.sec as u64,
+            entry_valid_nsec: ttl.nsec as u32,
+            attr_valid_nsec: ttl.nsec as u32,
             attr: fuse_attr_from_attr(attr),
         }, fuse_open_out {
             fh: fh,
